third_party/WebKit/LayoutTests/fast/css/fontfaceset-set-operations-expected.txt

Tests Set operations of FontFaceSet.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.fonts.size is 3
PASS callbackArgs[0] === callbackArgs[1] is true
PASS callbackArgs[2] === document.fonts is true
PASS thisValue === thisArg is true
PASS faces.length is 3
PASS faces[0].family is "Font1"
PASS faces[1].family is "Font2"
PASS faces[2].family is "Font3"
PASS document.fonts.has(faces[0]) is true
PASS document.fonts.has(faces[1]) is true
PASS document.fonts.has(faces[2]) is true
PASS document.fonts.has(new FontFace("MyFont", "local(Arial)", {})) is false
PASS document.fonts.has("Font1") threw exception TypeError: Failed to execute 'has' on 'FontFaceSet': The argument is not a FontFace..
PASS successfullyParsed is true

TEST COMPLETE
